ingestion-and-storage/requirements.txt
Faker==24.4.0
kafka-python==2.0.2
python-dotenv==1.0.1